@gvanrossum 2016-03-05
Should the @abstractmethod decorators also be removed? They're non-functional without ABCMeta.