@JukkaL 2016-03-05
Mypy can still check that abstract methods have been implemented in subclasses and that you don't instantiate an abstract class, so `@abstractmethod` without the ABCMeta is actually useful, though only during type checking as they will be ignored at runtime.