@gvanrossum 2016-03-05
Oh, I didn't realize mypy checks for @abstractmethod. That's great!
Will merge now.